requirements.txt
streamlit
requests
beautifulsoup4
sentence-transformers
numpy
pandas
groq
chromadb
python-dotenv
lxml